Rename modules and packages to better express intent
