+++ modules/models/canonical/dependencies.sbt
