+++ modules/models/canonical/metadata.sbt
@@ -1,0 +1,2 @@
+name := "models-canonical"
+description := "Strict representation of entities and component parts."
