+++ modules/models/common/metadata.sbt
@@ -1,0 +1,2 @@
+name := "models-common"
+description := "Concepts express less formally for use context outside of identity."
