+++ modules/models/common/src/main/scala/gig/dmf/models/common/Entity.scala
@@ -1,4 +1,4 @@
-package gig.dmf.core
+package gig.dmf.models.common
 
 import scala.collection.immutable._
