--- modules/persistence-cbor/metadata.sbt
@@ -1,2 +1,0 @@
-name := "core"
-description := "Concerns common across all modules."
