--- modules/persistence/.gitignore
@@ -1,1 +1,0 @@
-/target
