--- modules/persistence/compiler.sbt
@@ -1,13 +1,0 @@
-scalacOptions --=
-  //"-Wunused:implicits" ::
-  "-Wunused:imports" ::
-    //"-Wunused:locals" ::
-    //"-Wunused:params" ::
-    //"-Wunused:privates" ::
-    "-Ywarn-unused-import" ::
-    //"-Ywarn-unused:implicits" ::
-    "-Ywarn-unused:imports" ::
-    //"-Ywarn-unused:locals" ::
-    //"-Ywarn-unused:params" ::
-    //"-Ywarn-unused:privates" ::
-    Nil
