--- modules/persistence/metadata.sbt
@@ -1,2 +1,0 @@
-name := "persistence"
-description := "Storage concerns."
